requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
requests
tensorflow
pillow